+++ libs/ngx-material-intl-tel-input-lib/src/lib/validators/tel.validators.ts
@@ -1,7 +1,7 @@
 import { AbstractControl, FormGroup, ValidatorFn } from '@angular/forms';
 import { PhoneNumberFormat, PhoneNumberUtil } from 'google-libphonenumber';
 import { Country } from '../types/country.model';
-import { isValidPhoneNumberLength } from '../utils/phone-number.utils';
+import { analyzeTelephoneNumber } from './telephone-number.schema';
 
 export default class TelValidators {
   static isValidNumber(
@@ -69,23 +69,18 @@ export default class TelValidators {
           .get('numberControl')
           ?.setValue(formattedOnlyNumber, { emitEvent: false });
 
-        const isValidNumber = phoneNumberUtil.isValidNumber(parsed);
         setPrefixControlValue(parsed.getCountryCode(), allCountries, telForm);
 
-        // Check if the phone number length is valid for the country
         const countryIso = telForm?.value?.prefixCtrl?.iso2;
-        const isValidLength = countryIso
-          ? isValidPhoneNumberLength(control.value, countryIso)
-          : true;
+        const analysis = analyzeTelephoneNumber(control.value, countryIso);
 
-        if (!isValidNumber) {
+        if (analysis.error === 'invalidNumber') {
           control.setErrors({ invalidNumber: true });
           telForm.get('numberControl')?.setErrors({ invalidNumber: true });
           return {
             invalidNumber: true
           };
-        } else if (!isValidLength) {
-          // If the number is valid according to libphonenumber but exceeds the maximum length
+        } else if (analysis.error === 'numberTooLong') {
           control.setErrors({ numberTooLong: true });
           telForm.get('numberControl')?.setErrors({ numberTooLong: true });
           return {